sourceslist: add remove param to SourcesList.save()
Any files in the list that have all their entries removed from the list
will be completely ignored during save(), instead of overwriting and/or
removing the file.

The current SoftwareProperties code works around this problem by
adding an empty (invalid) SourceEntry whenever it detects that it
is removing the final SourceEntry for a file.

This does not change the default behavior, so the new action of removing
files will only happen for callers that set remove=True.