python: expose APT::Progress::PackageManagerFancy
In order to provide a better text based install progress this
commit exposes the APT::Progress::PackageManagerFancy class so
that it can be used with "pkgPackageManager.DoInstall()".

This is only a first step and the install progress should be
reworked in python-apt as it's quite different from the way APT
works. This is a first step on the way.